Enable typescript build.
