+++ config/webpack.plugins.js
@@ -39,13 +39,13 @@ const plugins = (dev = false) => [
     shared: [
       { react: { singleton: true, eager: true, requiredVersion: deps.react } },
       { 'react-dom': { singleton: true, eager: true, requiredVersion: deps['react-dom'] } },
-      { 'react-router-dom': { requiredVersion: deps['react-router-dom'] } },
-      { 'react-redux': { requiredVersion: deps['react-redux'] } },
-      { '@patternfly/react-table': { requiredVersion: deps['@patternfly/react-table'] } },
-      { '@patternfly/react-core': { requiredVersion: deps['@patternfly/react-core'] } },
-      { '@patternfly/quickstarts': { singleton: true, requiredVersion: deps['@patternfly/quickstarts'] } },
-      { '@scalprum/react-core': { singleton: true, requiredVersion: deps['@scalprum/react-core'] } },
-      { '@unleash/proxy-client-react': { singleton: true, requiredVersion: deps['@unleash/proxy-client-react'] } },
+      { 'react-router-dom': { requiredVersion: deps['react-router-dom'], eager: true } },
+      { 'react-redux': { requiredVersion: deps['react-redux'], eager: true } },
+      { '@patternfly/react-table': { requiredVersion: deps['@patternfly/react-table'], eager: true } },
+      { '@patternfly/react-core': { requiredVersion: deps['@patternfly/react-core'], eager: true } },
+      { '@patternfly/quickstarts': { singleton: true, requiredVersion: deps['@patternfly/quickstarts'], eager: true } },
+      { '@scalprum/react-core': { singleton: true, requiredVersion: deps['@scalprum/react-core'], eager: true } },
+      { '@unleash/proxy-client-react': { singleton: true, requiredVersion: deps['@unleash/proxy-client-react'], eager: true } },
     ],
   }),
   ChunkMapper,
