+++ package.json
@@ -126,6 +126,8 @@
     "stream-browserify": "^3.0.0",
     "style-loader": "^3.2.1",
     "terser-webpack-plugin": "^5.2.0",
+    "ts-loader": "^9.2.7",
+    "typescript": "^4.6.2",
     "url": "^0.11.0",
     "webpack": "^5.51.1",
     "webpack-bundle-analyzer": "^4.4.2",
